+++ src/__tests__/index.test.ts
@@ -1,23 +1,23 @@
-import { Fork, Sponsor, Star, UsedBy, Watch } from "../index";
+import { Fork, Sponsor, Star, UsedBy, Watch } from '../index';
 
-describe(">>> Exports required components", () => {
-  it("+++ Fork is truthy", () => {
+describe('>>> Exports required components', () => {
+  it('+++ Fork is truthy', () => {
     expect(Fork).toBeTruthy();
   });
 
-  it("+++ Sponsor is truthy", () => {
+  it('+++ Sponsor is truthy', () => {
     expect(Sponsor).toBeTruthy();
   });
 
-  it("+++ Star is truthy", () => {
+  it('+++ Star is truthy', () => {
     expect(Star).toBeTruthy();
   });
 
-  it("+++ UsedBy is truthy", () => {
+  it('+++ UsedBy is truthy', () => {
     expect(UsedBy).toBeTruthy();
   });
 
-  it("+++ Watch is truthy", () => {
+  it('+++ Watch is truthy', () => {
     expect(Watch).toBeTruthy();
   });
 });