+++ src/watch.tsx
@@ -1,10 +1,10 @@
 /**
  * @function Watch
  */
-import * as PropTypes from "prop-types";
-import * as React from "react";
+import * as PropTypes from 'prop-types';
+import * as React from 'react';
 
-import GithubButton from "./lib/githubButton";
+import GithubButton from './lib/githubButton';
 
 export interface IPropTypes {
   owner: string;
@@ -38,6 +38,6 @@ export default class Watch extends React.Component<IPropTypes, IState> {
   render() {
     // const { owner, repo } = this.props;
     const { watchers_count } = this.state;
-    return <GithubButton variant="watch" count={watchers_count} { ...this.props } />;
+    return <GithubButton variant="watch" count={watchers_count} {...this.props} />;
   }
 }